refactor: move posting
