+++ src/features/map-posting/services/posting-shapes.ts
